requirements.txt
APScheduler==3.5.1
asn1crypto==0.24.0
attrs==17.4.0
boto3==1.7.2
botocore==1.10.2
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
cryptography==2.2.2
cycler==0.10.0
docutils==0.14
flake8==3.5.0
flake8-polyfill==1.0.2
Flask==0.12.2
idna==2.6
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
lxml==4.2.1
mando==0.6.4
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
more-itertools==4.1.0
ndg-httpsclient==0.4.4
numpy==1.14.2
pandas==0.22.0
pluggy==0.6.0
py==1.5.3
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pytest==3.5.0
pytest-cov==2.5.1
pytest-flask==0.10.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
Quandl==3.3.0
radon==2.2.0
redis==2.10.6
requests==2.18.4
requests-cache==0.4.13
requests-file==1.4.3
requests-ftp==0.3.1
s3transfer==0.1.13
scipy==1.0.1
six==1.11.0
tzlocal==1.5.1
urllib3==1.22
weather-api==1.0.2
Werkzeug==0.14.1
wrapt==1.10.11
